add support for "unresolved" texture frames in texture registry
Up until now it was only allowed to push an
OpenGL frame with it's name, format etc, which
the engine will fetch from the external texture as
soon as possible.

Now, you can push an "unresolved" frame, which
is basically a callback and an opaque `void*`
pointer, which the engine texture registry will
call when the engine wants to know the exact
texture name, format, target etc, to fill in
("resolve") those details.

Basically allows you to be a bit smarter.
Maybe, in the time between
`texture_push_frame_unresolved` and the actual
execution of the resolve callback, a new
frame has arrived in the video player, and you
want to use that instead.