flutter-pi: make context current for engine run
FlutterEngineRunInitialized resolves the GL procs in the same thread
when using impeller, so an EGL context needs to be current. Make the
setup context current for the duration of FlutterEngineRunInitialized.

Also, this means that we can't use the setup context in the
`on_make_current` callback anymore, since that's called on the
rasterizer thread, possibly while `FlutterEngineRunInitialized` is
still running. Just use the rendering context there. (No real reason
to use the setup context anyway)