video player: fix for unsupported `EGL_EXT_..._modifiers`
If `EGL_EXT_image_dma_buf_import_modifiers` is not supported, we can't
query which pixel formats & modifiers are supported.
In that case, try creating an image with every pixel format (by
testing import with a small XRGB8888 gbm bo)

That hopefully gives a bit of an idea which pixel formats are working.
It's still technically UB because we're not importing with an explicit
DRM_FORMAT_MOD_LINEAR, so the driver could technically assume any
modifier it wants.