drmdev: call release callback before scanout callback
Call the release callback for the commit that is no longer visible on screen before calling the scanout callback.

The scanout callback might already start a new frame, and the release callback releases some resources that could be useful for building the new frame, so this helps avoid resource exhaustion.