WIP: feat: implement frame timing
Make frame_scheduler able to time frames, e.g. tell flutter when to
start rendering a frame when it receives a frame request.

For now, if no frame is currently waiting to be scanned out, this
just responds to the frame request with the current frame and a 60Hz
frame interval, i.e. starts the frame immediately.

If a frame is waiting to be scanned out, this waits until the scanout
to reply to the frame timings request. (With the timestamp of the
scanout being used as the frame start time in the reply to the frame
timings request)

If another frame is already queued after that, the scheduler will
wait until that other queued frame is scanned out to reply to the frame
timings request.