always force zero rotation in KMS planes for consistency
primary planes are sometimes rotated when the framebuffer console is
rotated. While it is potentially useful to respect that, and rotate the
app accordingly, it won't rotate the input of the app nor the cursor
plane.

Just use zero rotation everywhere.

Provide feedback if a cursor vs. overlay plane was allocated, if
prefer_cursor was set.

We can only use fast cursor updates if a cursor plane was allocated.